+++ service/v1/pendingTransactionsApi/getInvoiceInfo.js
@@ -54,12 +54,12 @@ module.exports = {
         amount: invoiceAmount
       })
       .then(decisionResult => {
-        invoiceDetails.fee = decisionResult.fee && decisionResult.fee.amount || 0
+        invoiceDetails.fee = (decisionResult.fee && decisionResult.fee.amount) || 0
         return invoiceDetails
         // return this.bus.importMethod('spsp.rule.decision.fetch')({
         //   currency: invoice.currencyCode,
         //   amount: invoiceAmount,
-        //   identifier: invoice.userNumber
+        //   identifier: invoice.identifier
         // })
         // .then(spspDecisionResult => {
         //   if (spspDecisionResult.sourceAmount) {